+++ app/src/main/java/com/flux/data/model/NotesModel.kt
@@ -7,11 +7,11 @@ import java.util.Date
 @Entity
 data class NotesModel(
     @PrimaryKey(autoGenerate = true)
-    val notesId: Int = 0,
+    val notesId: Long = 0L,
     val title: String = "",
     val description: String = "",
     val isPinned: Boolean = false,
-    val workspaceId: Int=0,
-    val labels: List<Int> = emptyList(),
+    val workspaceId: Long = 0L,
+    val labels: List<Long> = emptyList(),
     val lastEdited: Date = Date()
 )
