+++ app/src/main/java/com/flux/data/repository/TodoRepository.kt
@@ -6,6 +6,6 @@ import kotlinx.coroutines.flow.Flow
 interface TodoRepository {
     suspend fun upsertList(list: TodoModel)
     suspend fun deleteList(list: TodoModel)
-    suspend fun deleteAllWorkspaceLists(workspaceId: Int)
-    fun loadAllLists(workspaceId: Int): Flow<List<TodoModel>>
+    suspend fun deleteAllWorkspaceLists(workspaceId: Long)
+    fun loadAllLists(workspaceId: Long): Flow<List<TodoModel>>
 }
