+++ app/src/main/java/com/flux/data/repository/TodoRepositoryImpl.kt
@@ -12,8 +12,8 @@ class TodoRepositoryImpl @Inject constructor(
     val dao: TodoDao,
     @IODispatcher private val ioDispatcher: CoroutineDispatcher
 ): TodoRepository {
-    override fun loadAllLists(workspaceId: Int): Flow<List<TodoModel>> { return dao.loadAllLists(workspaceId) }
+    override fun loadAllLists(workspaceId: Long): Flow<List<TodoModel>> { return dao.loadAllLists(workspaceId) }
     override suspend fun upsertList(list: TodoModel) { return withContext(ioDispatcher) { dao.upsertList(list) } }
     override suspend fun deleteList(list: TodoModel) { return withContext(ioDispatcher) { dao.deleteList(list) } }
-    override suspend fun deleteAllWorkspaceLists(workspaceId: Int) { return withContext(ioDispatcher) { dao.deleteAllWorkspaceLists(workspaceId) } }
+    override suspend fun deleteAllWorkspaceLists(workspaceId: Long) { return withContext(ioDispatcher) { dao.deleteAllWorkspaceLists(workspaceId) } }
 }
