+++ app/src/main/java/com/flux/ui/events/NotesEvents.kt
@@ -4,9 +4,9 @@ import com.flux.data.model.LabelModel
 import com.flux.data.model.NotesModel
 
 sealed class NotesEvents {
-    data class DeleteAllWorkspaceNotes(val workspaceId: Int): NotesEvents()
-    data class LoadAllNotes(val workspaceId: Int): NotesEvents()
-    data class LoadAllLabels(val workspaceId: Int): NotesEvents()
+    data class DeleteAllWorkspaceNotes(val workspaceId: Long): NotesEvents()
+    data class LoadAllNotes(val workspaceId: Long): NotesEvents()
+    data class LoadAllLabels(val workspaceId: Long): NotesEvents()
     data class DeleteNote(val data: NotesModel): NotesEvents()
     data class DeleteNotes(val data: List<NotesModel>): NotesEvents()
     data class TogglePinMultiple(val data: List<NotesModel>): NotesEvents()
