+++ app/src/main/java/com/flux/ui/events/TodoEvents.kt
@@ -3,8 +3,8 @@ package com.flux.ui.events
 import com.flux.data.model.TodoModel
 
 sealed class TodoEvents {
-    data class DeleteAllWorkspaceLists(val workspaceId: Int): TodoEvents()
-    data class LoadAllLists(val workspaceId: Int): TodoEvents()
+    data class DeleteAllWorkspaceLists(val workspaceId: Long): TodoEvents()
+    data class LoadAllLists(val workspaceId: Long): TodoEvents()
     data class DeleteList(val data: TodoModel): TodoEvents()
     data class UpsertList(val data: TodoModel): TodoEvents()
 }
